requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
llama-cpp-python
numpy==1.26.3
torch
transformers==4.36.2
python-multipart==0.0.7
httpx==0.26.0
asyncio==3.4.3
pytest==7.4.3
pytest-asyncio==0.23.2
datasets
